+++ django-testbed/frontend/src/App.js
@@ -1,28 +1,15 @@
-import { Input } from 'antd';
+import { AntdUiProvider } from '@prestojs/ui-antd';
+import { mount, redirect, route } from 'navi';
 import React from 'react';
-import { UiProvider, getFormatterForField } from '@prestojs/ui';
-import {
-    FormWrapper,
-    getWidgetForField as antdGetWidgetForField,
-    FormItemWrapper,
-} from '@prestojs/ui-antd';
-import { SWRConfig } from 'swr';
 import { Router, View } from 'react-navi';
-import { mount, route, redirect } from 'navi';
+import { SWRConfig } from 'swr';
+
+import DatePicker from './DatePicker';
 
 // eslint-disable-next-line import/extensions
 import './styles/global.less?no-css-modules';
+import TimePicker from './TimePicker';
 import UserListView from './views/UserListView';
-
-const DefaultWidget = ({ input }) => <Input {...input} />;
-
-function getWidgetForField(field) {
-    const widget = antdGetWidgetForField(field);
-    if (!widget) {
-        return DefaultWidget;
-    }
-    return widget;
-}
 
 const routes = mount({
     '/': redirect('/users/'),
@@ -35,18 +22,13 @@ const routes = mount({
 export default function App() {
     return (
         <React.Suspense fallback={<div>Loading...</div>}>
-            <UiProvider
-                getWidgetForField={getWidgetForField}
-                getFormatterForField={getFormatterForField}
-                formItemComponent={FormItemWrapper}
-                formComponent={FormWrapper}
-            >
+            <AntdUiProvider datePickerComponent={DatePicker} timePickerComponent={TimePicker}>
                 <SWRConfig>
                     <Router routes={routes}>
                         <View />
                     </Router>
                 </SWRConfig>
-            </UiProvider>
+            </AntdUiProvider>
         </React.Suspense>
     );
 }