+++ js-packages/@prestojs/ui/src/getFormatterForField.ts
@@ -64,7 +64,7 @@ const choicesMapping = new Map<string, any>([
     ['IntegerField', ChoiceFormatter],
 ]);
 
-/*
+/**
  * Returns the default formatter for any given Field.
  *
  * Depending on Field, this will return either a Formatter component directly, or [Formatter, props] where props is the default props that would be applied to said formatter.